Actualización de cambios, finalización de los blueprints de public en el back, corección de un par de bugs, y desarrollo de template miscompras
