+++ api/admin/routes.py
@@ -1,5 +1,7 @@
 import mysql.connector
-from flask import Blueprint, request, jsonify, session
+from flask import Blueprint, abort, request, jsonify, session
+
+from db import get_connection
 
 admin_bp = Blueprint('admin', __name__)
 
@@ -175,60 +177,60 @@ def listar_categorias():
 
 ##############################################################################################
 
-@public_bp.route('/productos/<int:producto_id>', methods=['PATCH'])
-def api_update_stock(producto_id):
-    data = request.get_json() or {}
-    new_stock = data.get('stock')
-    if new_stock is None:
-        abort(400, 'stock requerido')
-    db = get_connection(); cur = db.cursor()
-    try:
-        cur.execute("UPDATE productos SET stock=%s WHERE id=%s", (new_stock, producto_id))
-        db.commit()
-        return jsonify({'msg':'Stock actualizado'}), 200
-    except mysql.connector.Error as e:
-        db.rollback()
-        abort(400, str(e))
+# @admin_bp.route('/productos/<int:producto_id>', methods=['PATCH'])
+# def api_update_stock(producto_id):
+#     data = request.get_json() or {}
+#     new_stock = data.get('stock')
+#     if new_stock is None:
+#         abort(400, 'stock requerido')
+#     db = get_connection(); cur = db.cursor()
+#     try:
+#         cur.execute("UPDATE productos SET stock=%s WHERE id=%s", (new_stock, producto_id))
+#         db.commit()
+#         return jsonify({'msg':'Stock actualizado'}), 200
+#     except mysql.connector.Error as e:
+#         db.rollback()
+#         abort(400, str(e))
 
 
 
-@app.route('/usuario/admin/modificar/<int:id>', methods=['GET', 'POST'])
-def modificar_producto(id):
-    coneccion = get_db()
-    cursor = coneccion.cursor(dictionary=True)
-    data = request.get_json()
-    print(data)
-    cursor.execute("SELECT * FROM productos WHERE id=%s;", (id,))
-    producto = cursor.fetchone()
-    if producto:
-        cursor.execute("UPDATE productos SET categoria = %s, nombre= %s, precio= %s, stock=%s, descripcion= %s, imagen=%s WHERE id = %s;", 
-                       (int(data["categoria"]), data["nombre"], float(data["precio"]), int(data["stock"]), data["descripcion"], data["imagen"], id,))
-    coneccion.commit()
-    cursor.close()
-    coneccion.close()
-    return jsonify({"message": "Producto modificado exitosamente"}), 201
+# @admin_bp.route('/usuario/admin/modificar/<int:id>', methods=['GET', 'POST'])
+# def modificar_producto(id):
+#     coneccion = get_db()
+#     cursor = coneccion.cursor(dictionary=True)
+#     data = request.get_json()
+#     print(data)
+#     cursor.execute("SELECT * FROM productos WHERE id=%s;", (id,))
+#     producto = cursor.fetchone()
+#     if producto:
+#         cursor.execute("UPDATE productos SET categoria = %s, nombre= %s, precio= %s, stock=%s, descripcion= %s, imagen=%s WHERE id = %s;", 
+#                        (int(data["categoria"]), data["nombre"], float(data["precio"]), int(data["stock"]), data["descripcion"], data["imagen"], id,))
+#     coneccion.commit()
+#     cursor.close()
+#     coneccion.close()
+#     return jsonify({"message": "Producto modificado exitosamente"}), 201
 
-@app.route('/usuario/admin/cargar', methods=['GET', 'POST'])
-def cargar():
-    coneccion = get_db()
-    cursor = coneccion.cursor(dictionary=True)
-    data = request.get_json()
-    cursor.execute("INSERT INTO productos (categoria, nombre, descripcion, precio, imagen, stock) VALUES (%s, %s, %s, %s, %s, %s);", 
-                       (int(data["categoria"]), data["nombre"], data["descripcion"], float(data["precio"]), data["imagen"],int(data["stock"]),))
-    coneccion.commit()
-    cursor.close()
-    coneccion.close()
-    return jsonify({"message": "Producto agregado exitosamente"}), 201
+# @admin_bp.route('/usuario/admin/cargar', methods=['GET', 'POST'])
+# def cargar():
+#     coneccion = get_db()
+#     cursor = coneccion.cursor(dictionary=True)
+#     data = request.get_json()
+#     cursor.execute("INSERT INTO productos (categoria, nombre, descripcion, precio, imagen, stock) VALUES (%s, %s, %s, %s, %s, %s);", 
+#                        (int(data["categoria"]), data["nombre"], data["descripcion"], float(data["precio"]), data["imagen"],int(data["stock"]),))
+#     coneccion.commit()
+#     cursor.close()
+#     coneccion.close()
+#     return jsonify({"message": "Producto agregado exitosamente"}), 201
 
 
-@app.route('/usuario/admin/borrar/<int:id>', methods=['DELETE'])
-def borrar(id):
-    coneccion = get_db()
-    cursor = coneccion.cursor(dictionary=True)
-    cursor.execute("DELETE FROM productos WHERE id = %s;", (id,))
-    if cursor.rowcount == 0:
-                return jsonify({"message": "Producto no encontrado"}), 404
-    coneccion.commit()
-    cursor.close()
-    coneccion.close()
-    return jsonify({"message": "Producto eliminado exitosamente"}), 200
+# @admin_bp.route('/usuario/admin/borrar/<int:id>', methods=['DELETE'])
+# def borrar(id):
+#     coneccion = get_db()
+#     cursor = coneccion.cursor(dictionary=True)
+#     cursor.execute("DELETE FROM productos WHERE id = %s;", (id,))
+#     if cursor.rowcount == 0:
+#                 return jsonify({"message": "Producto no encontrado"}), 404
+#     coneccion.commit()
+#     cursor.close()
+#     coneccion.close()
+#     return jsonify({"message": "Producto eliminado exitosamente"}), 200
