+++ api/app.py
@@ -3,6 +3,7 @@ from flask_cors import CORS
 from public.routes import public_bp
 from public.usuarios import usuarios_bp
 from public.productos import productos_bp
+from public.carrito import carrito_bp
 from admin.routes import admin_bp
 from public.miscompras import miscompras_bp
 
@@ -16,6 +17,7 @@ app.register_blueprint(productos_bp, url_prefix='/api/productos')
 app.register_blueprint(usuarios_bp, url_prefix='/api/usuarios')
 app.register_blueprint(admin_bp, url_prefix='/api/admin')
 app.register_blueprint(miscompras_bp, url_prefix='/api/miscompras')
+app.register_blueprint(carrito_bp, url_prefix='/api/carrito')
 
 if __name__ == '__main__':
     app.run(port=5001)
