+++ api/script_data_init.sql
@@ -3,24 +3,37 @@ VALUES
     ("")
 
 
-INSERT INTO productos (categoria, nombre, descripcion, precio, imagen, stock) VALUES
-(1, 'Catan', 'Juego de estrategia y comercio ambientado en una isla en expansión.', 1200, 'catan.jpg', 15),
-(2, 'Carcassonne', 'Juego de losetas para construir ciudades, caminos y monasterios.', 1000, 'carcassonne.jpg', 10),
-(3, 'Dixit', 'Cartas ilustradas donde la creatividad y la interpretación son claves.', 950, 'dixit.jpg', 12),
-(4, 'Ticket to Ride', 'Juego de trenes donde compites por rutas ferroviarias.', 1100, 'ticket_to_ride.jpg', 8),
-(5, 'Risk', 'Juego de conquista y estrategia militar a escala mundial.', 1150, 'risk.jpg', 20),
-(6, 'Monopoly', 'Clásico juego de propiedades, negocios y quiebras.', 900, 'monopoly.jpg', 25),
-(7, 'Aventureros al Tren: Europa', 'Versión europea de Ticket to Ride con estaciones y túneles.', 1200, 'aventureros_europa.jpg', 7),
-(8, '7 Wonders', 'Construye una civilización antigua a través de cartas.', 1300, '7wonders.jpg', 9),
-(1, 'Pandemic', 'Juego cooperativo para salvar al mundo de epidemias.', 1150, 'pandemic.jpg', 6),
-(2, 'Azul', 'Juego abstracto de colocación de azulejos portugueses.', 1050, 'azul.jpg', 13),
-(3, 'Splendor', 'Juego de comercio de gemas y desarrollo económico.', 950, 'splendor.jpg', 11),
-(4, 'Terraforming Mars', 'Transforma Marte en un planeta habitable.', 1600, 'terraforming_mars.jpg', 5),
-(5, 'Codenames', 'Juego de palabras con pistas para adivinar agentes secretos.', 850, 'codenames.jpg', 18),
-(6, 'Scythe', 'Juego de estrategia en una Europa alternativa postguerra.', 1800, 'scythe.jpg', 4),
-(7, 'King of Tokyo', 'Monstruos compiten por el dominio de Tokio.', 950, 'king_tokyo.jpg', 14),
-(8, 'Bang!', 'Juego del lejano oeste con roles ocultos y tiroteos.', 800, 'bang.jpg', 16),
-(1, 'Dominion', 'Juego de construcción de mazos ambientado en un reino.', 1200, 'dominion.jpg', 7),
-(2, 'The Mind', 'Juego cooperativo de sincronización sin comunicación.', 650, 'themind.jpg', 10),
-(3, 'Clue', 'Juego de deducción para encontrar al asesino.', 950, 'clue.jpg', 9),
-(4, 'Uno', 'Juego de cartas clásico y rápido para toda la familia.', 500, 'uno.jpg', 30);
+INSERT INTO productos (categoria, nombre, descripcion, precio, imagen, stock) 
+VALUES
+    (1, 'Catan', 'Juego de estrategia y comercio ambientado en una isla en expansión.', 1200, 'catan.jpg', 15),
+    (2, 'Carcassonne', 'Juego de losetas para construir ciudades, caminos y monasterios.', 1000, 'carcassonne.jpg', 10),
+    (3, 'Dixit', 'Cartas ilustradas donde la creatividad y la interpretación son claves.', 950, 'dixit.jpg', 12),
+    (4, 'Ticket to Ride', 'Juego de trenes donde compites por rutas ferroviarias.', 1100, 'ticket_to_ride.jpg', 8),
+    (5, 'Risk', 'Juego de conquista y estrategia militar a escala mundial.', 1150, 'risk.jpg', 20),
+    (6, 'Monopoly', 'Clásico juego de propiedades, negocios y quiebras.', 900, 'monopoly.jpg', 25),
+    (7, 'Aventureros al Tren: Europa', 'Versión europea de Ticket to Ride con estaciones y túneles.', 1200, 'aventureros_europa.jpg', 7),
+    (8, '7 Wonders', 'Construye una civilización antigua a través de cartas.', 1300, '7wonders.jpg', 9),
+    (1, 'Pandemic', 'Juego cooperativo para salvar al mundo de epidemias.', 1150, 'pandemic.jpg', 6),
+    (2, 'Azul', 'Juego abstracto de colocación de azulejos portugueses.', 1050, 'azul.jpg', 13),
+    (3, 'Splendor', 'Juego de comercio de gemas y desarrollo económico.', 950, 'splendor.jpg', 11),
+    (4, 'Terraforming Mars', 'Transforma Marte en un planeta habitable.', 1600, 'terraforming_mars.jpg', 5),
+    (5, 'Codenames', 'Juego de palabras con pistas para adivinar agentes secretos.', 850, 'codenames.jpg', 18),
+    (6, 'Scythe', 'Juego de estrategia en una Europa alternativa postguerra.', 1800, 'scythe.jpg', 4),
+    (7, 'King of Tokyo', 'Monstruos compiten por el dominio de Tokio.', 950, 'king_tokyo.jpg', 14),
+    (8, 'Bang!', 'Juego del lejano oeste con roles ocultos y tiroteos.', 800, 'bang.jpg', 16),
+    (1, 'Dominion', 'Juego de construcción de mazos ambientado en un reino.', 1200, 'dominion.jpg', 7),
+    (2, 'The Mind', 'Juego cooperativo de sincronización sin comunicación.', 650, 'themind.jpg', 10),
+    (3, 'Clue', 'Juego de deducción para encontrar al asesino.', 950, 'clue.jpg', 9),
+    (4, 'Uno', 'Juego de cartas clásico y rápido para toda la familia.', 500, 'uno.jpg', 30);
+
+
+INSERT INTO categoria (nombre) 
+VALUES
+    ("Estrategia"), 
+    ("Azar"), 
+    ("Rol"), 
+    ("Cartas"), 
+    ("Habilidad"), 
+    ("Cooperativos"), 
+    ("Solitario"), 
+    ("Adultos");
