+++ app/static/css/miscompras.css
@@ -1,6 +1,7 @@
 #miscompras-body {
     height: auto;
-    width: 100vw;
+    width: auto;
+    overflow: hidden;
 }
 
 :root{
@@ -12,57 +13,70 @@
     display: flex;
     align-items: center;
     height: 10vh;
-    width: 100vw;
+    width: auto;
     background-color: var(--grisesito);
     margin-bottom: 1px;
 }
 
 #miscompras-titulo {
+    font-size: 2.5rem;
     margin-top: 1.4vh;
     margin-left: 4.5vw;
 }
 
-#miscompras-content {
+#compras-content {
     display: flex;
     flex-direction: column;
     align-items: center;
-    min-height: 80vh;
-    height: auto;
+    height: fit-content;
     width: 100%;
 }
 
+#compra-content {
+    margin: 4rem;
+    width: 85%;
+    height: fit-content;
+    background-color: rgb(254, 254, 176);
+    display:block;
+    justify-content: center;
+    align-items: center;
+    border-radius: 1rem;
+    overflow: hidden;
+    border: solid black 1px;
+}
+
+#compra-id-wrapper {
+    background-color: var(--grisesitoclaro);
+    width: 100%;
+    height: 3.5rem;
+    display: flex;
+    align-items: center;
+    border-bottom: solid black 0.5px;
+}
+
+#compra-id-texto{
+    font-size: 2rem;
+    margin: 2px 1rem;
+}
+
 #miscompras-barradecompra {
-    margin-top: 3%;
+    margin: 2rem auto 2rem auto;
     background-color: rgb(235, 241, 240);
-    height: 34%;
-    width: 80%;
+    height: 14rem;
+    width: 85%;
     border: solid black 1.5px;
     overflow: hidden;
     border-radius: 0.5em;
 }
 
-#compra-fechadecompra-wrapper {
-    background-color: var(--grisesitoclaro);
-    width: 100%;
-    height: 17%;
-    display: flex;
-    align-items: baseline;
-    border-bottom: solid black 0.5px;
-}
-
-#compra-fechadecompra {
-    margin: 0.4rem;
-    margin-left: 1.2rem;
-}
-
 #compra-datos {
     display: flex;
     width: 100%;
-    height: 83%;
+    height: fit-content;
 }
 
 #compra-fotoproducto-wrapper {
-    margin-left: 3%;
+    margin: 1rem;
     background-color: white;
     height: 100%;
     width: 20%;
@@ -85,33 +99,17 @@
 }
 
 #compra-detallesproducto {
-
     margin-left: 2%;
     height: 100%;
-    width: 30%;
+    width: 40%;
 }
 
 .detalleproducto {
-    margin-top: 1em;
-    margin-bottom: 1em;
-}
-
-#compra-detallescompra {
-    height: 100%;
-    width: 20%;
+    margin: 2rem 0rem 2rem 1rem;
 }
 
 .detallecompra {
-    margin-top: 0.8em;
-    margin-bottom: 0.8em;
-}
-
-#opciones {
-    margin-top: 1em;
-    margin-left: 2%;
-    height: 80%;
-    width: 22%;
-    display:flexbox;
+    margin: 2rem 1rem 0rem 1rem;
 }
 
 .boton-ver-wrapper {
@@ -122,16 +120,14 @@
     align-items: center;
 }
 
-
-
-
-#boton-verproducto-container, #boton-vercompra-container{
+#boton-verproducto-container{
     display: flex;
     justify-content: center;
     align-items: center;
 }
 
 .opcionesvermas{
+    color: black;
     margin-top: 0.8em;
     padding: 0.4em;
     border: solid black 2px;
@@ -148,4 +144,8 @@
     -webkit-transition: all 200ms ease;
     -o-transition: all 200ms ease;
     transition: all 200ms ease;
+}
+
+h5{
+    text-align: justify;
 }
